Content Edits
1) We need the header Showcase Information just below the Our Goal
statement (as in the storyboard).

2) Delete the text "at the EDGE Stage" in the Our Goal section from the
sentence "6 finalists (1 regional, 1 Canadian, and 4 national) will
then perform live at the Festival at the EDGE Stage on Saturday, July
25th in front of a panel of judges and a to-be-announced Kingdom Bound
Artist."

3) Delete the text "on the EDGE Stage" in the Prizes section from the
sentence "The opportunity to perform live on The Edge Stage at the 2015
Kingdom Bound Showcase"

4) Add the text "and a $25 submission fee" to the sentence "We'll also
need two songs of your choice submitted as MP3s." in the Submission
Requirements section.

5) On the registration page the sentence in the red box should read as
follows "Upon submission, you will be redirected to Paypal to submit
your nonrefundable $25 registration fee. Failure to pay the
registration fee will result in a forfeit in your application, and will
not be considered for entry."